Moved LxControlState from Controls package to Types package and renamed to LxState
